# Conflicts:
#	core/api/make.cpp
#	core/api/make.hpp
#	core/api/node_api.hpp
#	core/api/rpc/ws.cpp
#	core/api/visit.hpp

+++ core/primitives/sector_file/sector_file.hpp
@@ -24,7 +24,7 @@ using fc::primitives::sector::RegisteredSealProof;
 using fc::primitives::sector::SectorId;
 
 namespace fc::primitives::sector_file {
-  enum SectorFileType : int {
+  enum SectorFileType : int64_t {
     FTNone = 0,
     FTUnsealed = 1,
     FTSealed = 2,
